revrited PAGES constant and revrite map for page side menu, remove Custom Menu component
