+++ src/app/components/index.js
@@ -1,5 +1,4 @@
 import PageWrapper from './PageWrapper'
 import Header from './Header'
-import CustomMenu from './Menu'
 import Dashboard from './Dashboard'
-export { PageWrapper, Header, CustomMenu, Dashboard }
+export { PageWrapper, Header, Dashboard }
