+++ src/app/constants/pages.js
@@ -1,11 +1,11 @@
 import { ROUTES_PATHS } from 'app/constants/'
 
-const PAGES = {
-  'Company dashboard': ROUTES_PATHS.COMPANIES_SHOW,
-  Calendar: ROUTES_PATHS.CALENDAR_SHOW,
-  Projects: ROUTES_PATHS.PROJECTS_ALL,
-  Reports: ROUTES_PATHS.REPORTS_ALL,
-  Statistic: ROUTES_PATHS.STATISTICS_ALL
-}
+const PAGES = [
+  { text: 'Company dashboard', path: ROUTES_PATHS.COMPANIES_SHOW },
+  { text: 'Calendar', path: ROUTES_PATHS.CALENDAR_SHOW },
+  { text: 'Projects', path: ROUTES_PATHS.PROJECTS_ALL },
+  { text: 'Reports', path: ROUTES_PATHS.REPORTS_ALL },
+  { text: 'Statistic', path: ROUTES_PATHS.STATISTICS_ALL }
+]
 
 export default PAGES
